sussa-nlp/requirements.txt
autopep8==2.0.2
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
easygoogletranslate==0.0.4
filelock==3.12.0
Flask==2.3.2
fsspec==2023.4.0
huggingface-hub==0.14.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.2.1
networkx==3.0
numpy==1.24.3
packaging==23.1
pycodestyle==2.10.0
python-environ==0.4.54
PyYAML==6.0
regex==2023.5.4
requests==2.29.0
sympy==1.11.1
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.3
